requirements.txt
opencv-python
pillow-avif-plugin
ultralytics
transformers
torch
Pillow
numpy
git-lfs
yolov5